refactor(.ajaja): infra 패키지 의존 관계 분리
